Stripped out some history code, and condensed functionality.
This is intened to be an interface to the weather station only, so removing a lot of code that's storing historical data.  I need to re-add a bit of code to find (which sensor to use?) an outdoor sensor and use that to calculate windchill, etc.

Original code is being migrated to LexacyCode until I can verify it's working at which point I will remove the Legacy Code.