A bit of refactoring to clean up code
Broke out an Enum to hold device specific information.  Changed stationList to an ArrayList for easier manipulation, broke out an append medthod for data, added a prepend to re-use data on oversized packets 

(Occasionally we get oversize valid packets, that we truncated and process, but the date trimmed off was getting lost, putting it back into the buffer in hopes it may be useful.