A bit more refactoring to split out buffer handling
Moved buffer handling and bit reading within the buffer out to it's own class, cleaned up all the formula's to use new methods in that class (thank God for tests!) making the formula's much more readable, and maintainable.  Updated rainfall to be in decimal inches rather than 1/10th of an inch (easier t read for people).    Added metric conversion on in to mm as well.